Added door checker and Firebase status updater to embedded code
Right now there is a simple three state SM that updates the Firebase DB
based on the machine movement and door status.  If the machine starts
moving, the SM goes into a washing state.  If it stops, it goes into a
waiting to be emptied state. Then if the door opens, it goes back to the
idle state.