Embedded code used to gather data for a full washing cycle.
Code has serial prints to print out data values to the screen.  First
column is unfiltered square of difference of x acceleration, second
column is unfiltered square of difference of y acceleration, third
column is unfiltered square of difference of z acceleration, and final
col is the magnitude of the three with a software lowpass.